requirements.txt
# Web框架和相关扩展
Flask==2.0.1
Flask-Session==0.4.0
Werkzeug==2.0.1

# 数据库
mysql-connector-python==8.0.26

# AI和自然语言处理
openai==1.3.0
dashscope==1.10.0  # 阿里云通义千问API

# RSS和网页处理
feedparser==6.0.8
beautifulsoup4==4.9.3
requests==2.26.0

# 邮件处理
email-validator==1.1.3

# 云服务
oss2==2.16.0  # 阿里云OSS SDK
aliyun-fc2==2.6.1  # 阿里云函数计算SDK

# 安全相关
python-dotenv==0.19.0
cryptography==3.4.7
PyJWT==2.1.0
bcrypt==3.2.0

# 工具库
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
urllib3==1.26.7

# 图表生成
mermaid-py==0.1.1

# 开发工具
pytest==6.2.5  # 测试框架
black==21.6b0  # 代码格式化
flake8==3.9.2  # 代码检查